rm deprecated execCommand, clean up clipboard copy fn
