+++ apps/portal/src/utils/copy-text-to-clipboard.js
@@ -1,0 +1,8 @@
+export async function copyTextToClipboard(text) {
+    try {
+        await window.navigator.clipboard.writeText(text);
+        return true;
+    } catch {
+        return false;
+    }
+}
